refactor: update js-sandbox directory hierarchy
