+++ packages/hoppscotch-js-sandbox/node.d.ts
@@ -1,2 +1,2 @@
-export { default } from "./dist/node.d.ts"
-export * from "./dist/node.d.ts"
+export { default } from "./dist/node/index.d.ts"
+export * from "./dist/node/index.d.ts"
