+++ packages/hoppscotch-js-sandbox/src/__tests__/preRequest.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test } from "vitest"
 
-import { runPreRequestScript } from "~/pre-request/node-vm"
+import { runPreRequestScript } from "~/node"
 
 describe("runPreRequestScript", () => {
   test("returns the updated environment properly", () => {
