--- packages/hoppscotch-js-sandbox/src/node.ts
@@ -1,2 +1,0 @@
-export * from "./pre-request/node-vm"
-export * from "./test-runner/node-vm"
