+++ packages/hoppscotch-js-sandbox/src/node/index.ts
@@ -1,0 +1,2 @@
+export { runPreRequestScript } from "./pre-request"
+export { runTestScript } from "./test-runner"
