+++ packages/hoppscotch-js-sandbox/src/node/test-runner.ts
@@ -7,28 +7,10 @@ import type ivmT from "isolated-vm"
 
 import { TestResponse, TestResult } from "~/types"
 import { getTestRunnerScriptMethods, preventCyclicObjects } from "~/utils"
+import { getSerializedAPIMethods } from "./utils"
 
 const nodeRequire = createRequire(import.meta.url)
 const ivm = nodeRequire("isolated-vm")
-
-// Function to recursively wrap functions in `ivm.Reference`
-const getSerializedAPIMethods = (
-  namespaceObj: Record<string, unknown>
-): Record<string, unknown> => {
-  const result: Record<string, unknown> = {}
-
-  for (const [key, value] of Object.entries(namespaceObj)) {
-    if (typeof value === "object" && value !== null && !Array.isArray(value)) {
-      result[key] = getSerializedAPIMethods(value as Record<string, unknown>)
-    } else if (typeof value === "function") {
-      result[key] = new ivm.Reference(value)
-    } else {
-      result[key] = value
-    }
-  }
-
-  return result
-}
 
 export const runTestScript = (
   testScript: string,
