--- packages/hoppscotch-js-sandbox/src/web.ts
@@ -1,2 +1,0 @@
-export * from "./pre-request/web-worker"
-export * from "./test-runner/web-worker"
