+++ packages/hoppscotch-js-sandbox/src/web/pre-request/worker.ts
