+++ packages/hoppscotch-js-sandbox/vite.config.ts
@@ -7,8 +7,8 @@ export default defineConfig({
     emptyOutDir: true,
     lib: {
       entry: {
-        web: "./src/web.ts",
-        node: "./src/node.ts",
+        web: "./src/web/index.ts",
+        node: "./src/node/index.ts",
       },
       name: "js-sandbox",
       formats: ["es", "cjs"],
